Move common acceptance methods to `plugin_test`
